Move vector_pack utilities to algorithms module
