+++ libs/pika/algorithms/include/pika/parallel/datapar/find.hpp
@@ -10,14 +10,14 @@
 
 #if defined(PIKA_HAVE_DATAPAR)
 #include <pika/execution/traits/is_execution_policy.hpp>
-#include <pika/execution/traits/vector_pack_all_any_none.hpp>
-#include <pika/execution/traits/vector_pack_find.hpp>
 #include <pika/functional/tag_invoke.hpp>
 #include <pika/parallel/algorithms/detail/find.hpp>
 #include <pika/parallel/datapar/iterator_helpers.hpp>
 #include <pika/parallel/datapar/loop.hpp>
 #include <pika/parallel/datapar/zip_iterator.hpp>
 #include <pika/parallel/util/result_types.hpp>
+#include <pika/parallel/util/vector_pack_all_any_none.hpp>
+#include <pika/parallel/util/vector_pack_find.hpp>
 
 #include <cstddef>
 #include <type_traits>
