+++ libs/pika/algorithms/include/pika/parallel/datapar/loop.hpp
@@ -10,15 +10,15 @@
 
 #if defined(PIKA_HAVE_DATAPAR)
 #include <pika/execution/traits/is_execution_policy.hpp>
-#include <pika/execution/traits/vector_pack_alignment_size.hpp>
-#include <pika/execution/traits/vector_pack_load_store.hpp>
-#include <pika/execution/traits/vector_pack_type.hpp>
 #include <pika/executors/datapar/execution_policy.hpp>
 #include <pika/executors/execution_policy.hpp>
 #include <pika/functional/tag_invoke.hpp>
 #include <pika/parallel/algorithms/detail/predicates.hpp>
 #include <pika/parallel/datapar/iterator_helpers.hpp>
 #include <pika/parallel/util/loop.hpp>
+#include <pika/parallel/util/vector_pack_alignment_size.hpp>
+#include <pika/parallel/util/vector_pack_load_store.hpp>
+#include <pika/parallel/util/vector_pack_type.hpp>
 
 #include <algorithm>
 #include <cstddef>
