+++ libs/pika/algorithms/include/pika/parallel/util/detail/simd/vector_pack_count_bits.hpp
