+++ libs/pika/algorithms/include/pika/parallel/util/vector_pack_alignment_size.hpp
@@ -67,7 +67,7 @@ namespace pika::parallel::traits::detail {
 }    // namespace pika::parallel::traits::detail
 
 #if !defined(__CUDACC__)
-#include <pika/execution/traits/detail/simd/vector_pack_alignment_size.hpp>
+#include <pika/parallel/util/detail/simd/vector_pack_alignment_size.hpp>
 #endif
 
 #endif
