+++ libs/pika/algorithms/include/pika/parallel/util/vector_pack_count_bits.hpp
@@ -21,7 +21,7 @@ namespace pika::parallel::traits::detail {
 #if defined(PIKA_HAVE_DATAPAR)
 
 #if !defined(__CUDACC__)
-#include <pika/execution/traits/detail/simd/vector_pack_count_bits.hpp>
+#include <pika/parallel/util/detail/simd/vector_pack_count_bits.hpp>
 #endif
 
 #endif
